Return files in subdirs from `tut`
Fixes #100.

The `tut` command does successfully recurse into subdirectories and
compile nested files. However, the List that it returns doesn't include
files that are within subdirectories. This causes nested files to not be
picked up by sbt-site if you do something like
`site.addMappingsToSiteDir(tut, "tut")`.

This change adds the touched nested files to the output of the command.
Since the result is no longer necessarily a flat directory, we have to
switch from using `File.getName` to a relative path for the file name
part of the (file, filename) tuple. I'm using `toString` on the relative
path which works just fine on my Mac, but I don't know if that could
cause trouble on Windows machines that might use a different path
delimiter.