Docs: scalacOptions for unused imports
I had an issue where I had a bunch of warnings when running tut with
the Tpolecat's scalaOptions! I took the options from:
https://tpolecat.github.io/2017/04/25/scalac-flags.html

Fortunately, in the FAQ, I could find a way to work around that. But
the flag for unused import in the FAQ is different from the one I had
in my options: `-Ywarn-unused-import` vs `-Ywarn-unused:imports`

It took me way too long to figure that out, so here it is for others like
me.